Throws error if .env does not exists.

+++ run.sh
@@ -1,4 +1,9 @@
 #!/bin/bash -e
+if [ ! -f ".env" ]; then
+  echo ".env file must exist. Please create it based on .env.tpl."
+  exit 1
+fi
+
 source .env
 docker-compose down
 docker-compose up
